refactor top-level inputs
nixdoc was passed as a source rather than a derivation.

I briefly considered passing `src` as a package recipe argument, but
there actually exists a package with that name. Using `sources` seems
the least worst compromise for tie time being. But the underlying
problem remains that it's not possible to express coarse-grained
references to remote sources within self-contained expressions, and
preserve the pins somewhere else for re-use:
https://github.com/fricklerhandwerk/module-sources